test new version 1

+++ src/train.py
@@ -1,33 +1,86 @@
 import sys
 from pathlib import Path
 from typing import Tuple
-
+import json
+import os
 import numpy as np
 import tensorflow as tf
 import yaml
-
+import repository as rs
 from utils.seed import set_seed
-def get_model_from_config(image_shape: Tuple[int, int, int], config: dict) -> tf.keras.Model:
+
+
+def get_model_from_config(image_shape: Tuple[int, int, int], config: dict,seed=1234) -> tf.keras.Model:
     """Create a CNN model based on YAML configuration."""
-
     key_config,val_config = next(iter(config.items()))
-
+    inputs = tf.keras.Input(shape=image_shape)
     if key_config == "ResNet50":
-        model=tf.keras.applications.ResNet50(
-            include_top=True,
-            weights=None,
+        base_model=tf.keras.applications.ResNet50(
+            include_top=False,
+            weights="imagenet",
             input_shape=image_shape,
-            classes=val_config,
         )
+        base_model.trainable = True
+        x = base_model(inputs, training=True)
+        x = tf.keras.layers.GlobalAveragePooling2D()(x)
+        x = tf.keras.layers.Dense(512, activation='relu')(x)
+        x = tf.keras.layers.BatchNormalization()(x)
+        x = tf.keras.layers.Dropout(0.5)(x)
+        outputs = tf.keras.layers.Dense(val_config, activation='softmax')(x)
     elif key_config == "MobileNet":
-        model=tf.keras.applications.MobileNet(
-            include_top=True,
+        base_model=tf.keras.applications.MobileNet(
+            include_top=False,
+            weights="imagenet",
+            input_shape=image_shape,
+        )
+        base_model.trainable = True
+        x = base_model(inputs, training=True)
+        x = tf.keras.layers.GlobalAveragePooling2D()(x)
+        x = tf.keras.layers.Dense(512, activation='relu')(x)
+        x = tf.keras.layers.BatchNormalization()(x)
+        x = tf.keras.layers.Dropout(0.5)(x)
+        outputs = tf.keras.layers.Dense(val_config, activation='softmax')(x)
+    elif key_config == "NASNetMobile":
+        base_model=tf.keras.applications.NASNetMobile(
+            include_top=False,
+            weights="imagenet",
+            input_shape=image_shape,
+        )
+        base_model.trainable = True
+        x = base_model(inputs, training=True)
+        x = tf.keras.layers.GlobalAveragePooling2D()(x)
+        x = tf.keras.layers.Dense(512, activation='relu')(x)
+        x = tf.keras.layers.BatchNormalization()(x)
+        x = tf.keras.layers.Dropout(0.5)(x)
+        outputs = tf.keras.layers.Dense(val_config, activation='softmax')(x)
+    elif key_config == "EfficientNetB0":
+        base_model=tf.keras.applications.EfficientNetB0(
+            include_top=False,
+            weights="imagenet",
+            input_shape=image_shape,
+        )
+        base_model.trainable = True
+        x = base_model(inputs, training=True)
+        x = tf.keras.layers.GlobalAveragePooling2D()(x)
+        x = tf.keras.layers.Dense(512, activation='relu')(x)
+        x = tf.keras.layers.BatchNormalization()(x)
+        x = tf.keras.layers.Dropout(0.5)(x)
+        outputs = tf.keras.layers.Dense(val_config, activation='softmax')(x)
+    elif key_config == "EfficientNetB1":
+        model=tf.keras.applications.EfficientNetB1(
+            include_top=False,
             weights=None,
             input_shape=image_shape,
             classes=val_config,
         )
+        base_model.trainable = True
+        x = base_model(inputs, training=True)
+        x = tf.keras.layers.GlobalAveragePooling2D()(x)
+        x = tf.keras.layers.Dense(512, activation='relu')(x)
+        x = tf.keras.layers.BatchNormalization()(x)
+        x = tf.keras.layers.Dropout(0.5)(x)
+        outputs = tf.keras.layers.Dense(val_config, activation='softmax')(x)
     else:
-        inputs = tf.keras.Input(shape=image_shape)
         x = inputs
         model = tf.keras.models.Sequential()
         for layer in config["layer"]:
@@ -41,6 +94,7 @@ def get_model_from_config(image_shape: Tuple[int, int, int], config: dict) -> tf
                             filters=branche_type[1]["filters"],
                             kernel_size=tuple(branche_type[1]["kernel_size"]),
                             activation=branche_type[1]["activation"],
+                            padding=branche_type[1]["padding"],
                         )(branch)
                     elif branche_type[0] == "max_pool":
                         branch = tf.keras.layers.MaxPooling2D(
@@ -48,6 +102,10 @@ def get_model_from_config(image_shape: Tuple[int, int, int], config: dict) -> tf
                             )(branch)
                     elif branche_type[0] == "flatten":
                         branch = tf.keras.layers.Flatten()(branch)
+                    elif branche_type[0] == "dropout":
+                        branch = tf.keras.layers.Dropout(branche_type[1], noise_shape=None, seed=seed,)(branch)
+                    elif branche_type[0] == "batch_norm":
+                        branch = tf.keras.layers.BatchNormalization()(branch)
                     elif branche_type[0] == "dense_layers":
                         branch = tf.keras.layers.Dense(
                             units=branche_type[1]["units"],
@@ -56,15 +114,35 @@ def get_model_from_config(image_shape: Tuple[int, int, int], config: dict) -> tf
                     elif branche_type[0] == "output_classes":
                         branch = tf.keras.layers.Dense(
                             units=branche_type[1],
-                            activation=None,  # Use softmax during compilation
+                            activation=None,  
                         )(branch)
             branches.append(branch)
             if len(branches)> 1:
                 x = tf.keras.layers.Concatenate()(branches)
             else:
                 x = branches[0]
-        model = tf.keras.Model(inputs=inputs, outputs=x)
+    model = tf.keras.Model(inputs=inputs, outputs=x)
     return model
+
+
+
+def check_model_exist(model,params):
+    """
+    check if a model is alrady train
+    if true reture model and metadata
+    if false return none
+    """
+    root_dir="model"
+    list_model_path=root_dir+"/list_model.json"
+    if os.path.exists(list_model_path):
+        if os.path.getsize(list_model_path) != 0:
+            with open(list_model_path, "r", encoding="utf-8") as file:
+                model_list = json.load(file)
+            repo = rs.RepositoryModel()
+            return repo.comp_list_model(model,params,model_list)
+        return None,None
+        
+
 
 def main() -> None:
     if len(sys.argv) != 3:
@@ -104,6 +182,10 @@ def main() -> None:
             lr = model_configs[model_v]["params"]["lr"]
         elif "epochs" in model_configs[model_v]["params"]:
             epochs = model_configs[model_v]["params"]["epochs"]
+    
+    param={"seed":seed,"lr":lr,"epochs":epochs}
+
+    
 
 
     # Set seed for reproducibility
@@ -112,6 +194,8 @@ def main() -> None:
     # Load data
     ds_train = tf.data.Dataset.load(str(prepared_dataset_folder / "train"))
     ds_val = tf.data.Dataset.load(str(prepared_dataset_folder / "val"))
+
+
     # Define the model
 
     #model = get_model(image_shape, conv_size, dense_size, output_classes,model_v)
@@ -125,21 +209,45 @@ def main() -> None:
     model.summary()
 
 
-    # Train the model
-    model.fit(
-        ds_train,
-        epochs=epochs,
-        validation_data=ds_val,
-    )
+    existing_model,metadata=check_model_exist(model=model,params=param)
 
-    # Save the model
-    model_folder.mkdir(parents=True, exist_ok=True)
-    model_path = model_folder / "model.keras"
-    model.save(model_path)
-    # Save the model history
-    np.save(model_folder / "history.npy", model.history.history)
+    #check if existe
+    if existing_model ==None:
+        # Train the model
+        model.fit(
+            ds_train,
+            epochs=epochs,
+            validation_data=ds_val,
+        )
+        #save model bento
+        repr = rs.RepositoryModel()
+        bento_model=repr.save_model(
+            name="bento-model"
+            ,model=model,
+            metadata={
+                "seed" : seed,
+                "lr": lr,
+                "epochs": epochs,
+            }
+        )
+        repr.export_model("bento-model")
+        # Save the model
+        model_folder.mkdir(parents=True, exist_ok=True)
+        model_path = model_folder / "model.keras"
+        model.save(model_path)
+        # Save the model history
+        np.save(model_folder / "history.npy", model.history.history)
 
-    print(f"\nModel saved at {model_folder.absolute()}")
+        with open(model_folder / "name_model.json", "w") as f:
+            json.dump({"name_model": str(bento_model.tag)}, f)
+        print(f"\nModel saved at {model_folder.absolute()}")
+    else:
+        #juste save model and history
+        model_folder.mkdir(parents=True, exist_ok=True)
+        model_path = model_folder / "model.keras"
+        existing_model.save(model_path)
+        model_folder.mkdir(parents=True, exist_ok=True)
+        np.save(model_folder / "history.npy", metadata)
 
 
 if __name__ == "__main__":
